Moves TryGetTestProperty to Futiliities.Testing as extension method and TestingBase class
