+++ Futilities/Futilities.ObjectExtensions.Tests/ObjectExtensionsTests.cs
@@ -1,3 +1,4 @@
+using Futilities.Testing;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using System;
 using System.Collections.Generic;
@@ -8,7 +9,7 @@ using System.Reflection;
 namespace Futilities.ObjectExtensions.Tests
 {
     [TestClass]
-    public class ObjectExtensionsTests
+    public class ObjectExtensionsTests : TestingBase
     {
         [TestMethod]
         [TestProperty("FooPropertyValue", "bar")]
@@ -36,31 +37,5 @@ namespace Futilities.ObjectExtensions.Tests
             Assert.AreEqual(testObject.Foo, testPropertyValue);
         }
 
-        [ExcludeFromCodeCoverage]
-        private bool TryGetTestProperty<T>(string methodName, string propertyName, out T propertyValue)
-        {
-            try
-            {
-                Type type = GetType();
-
-                MethodInfo methodInfo = type.GetMethod(methodName);
-
-                IEnumerable<TestPropertyAttribute> attribute = methodInfo.GetCustomAttributes<TestPropertyAttribute>();
-
-                var value = attribute.FirstOrDefault(a => a.Name == propertyName)?.Value ?? null;
-
-                propertyValue = (T)Convert.ChangeType(value, typeof(T));
-
-                if (value is null)
-                    return false;
-
-                return true;
-            }
-            catch (Exception)
-            {
-                propertyValue = default;
-                return false;
-            }
-        }
     }
 }
